rbac:refactor to keep cmd/*/ directory only with main.go
